refactor: improve fzf saveas functionality
Instead of adding a "new element" item to the fzf list we can save to a
new file by typing the name for our destination: If there is none
matching element it will automatically be saved to a new file. If there
is a matching file there will be another option to save it to the exact
query text instead. As a consequence the transform function option was
removed.